# openapi.yaml
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/creat":{"post":{"tags":["Создание и обработка заявок.(изменение)"],"summary":"Создание обращение","description":"Получение данных для создание заявки\nОтдает создали или нет","operationId":"creatPoint","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PointRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PointResponse"}}}}}}},"/api/all":{"get":{"tags":["Создание и обработка заявок.(изменение)"],"operationId":"allPointResponse","parameters":[{"name":"coordinates","in":"query","required":true,"schema":{"$ref":"#/components/schemas/AllPointResponse"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AllPointResponse"}}}}}}}},"components":{"schemas":{"PointRequest":{"type":"object","properties":{"pointCoordinates":{"type":"string","description":"координаты в формате 60.497874,56.926760 "},"description":{"type":"string","description":"Описание проблемы"}},"description":"Входящий запрос с координатами точки + описание проблемы + фото"},"PointResponse":{"type":"object","properties":{"status":{"type":"string","description":"статус обращения"},"pointCoordinates":{"type":"string","description":"координаты в формате 60.497874,56.926760 "},"description":{"type":"string","description":"Описание проблемы"}},"description":"Ответ API включает себя статус обращения,координаты,описание проблемы"},"AllPointResponse":{"type":"object","properties":{"points":{"type":"array","items":{"$ref":"#/components/schemas/PointResponse"}},"limit":{"type":"integer","format":"int32"},"offset":{"type":"integer","format":"int32"}},"description":"Список всех обращений"}}}}